# code/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Mock Server API
  version: 0.0.0
  description: |
    The control API of Mock Server: everything the panel can do, available to a
    script, a test suite or another application.

    It exists for the case where the mocks are not driven by hand. A test run
    picks which mock answers, exercises the system under test, and then asks the
    server what was actually called. All of that happens over this API, with no
    browser involved.

    ## The loop this API is built for

    1. `POST /routes` (or import them once from the panel) to have both variants
       of an endpoint: the happy path and the failure you want to rehearse.
    2. Tag them, for example `checkout-ok` and `checkout-down`.
    3. Before each case, `POST /routes/activate` with that tag and
       `exclusive: true`. The chosen mocks come up and the ones that competed
       for the same request go down, so there is no ambiguity about who answers.
    4. Run the test.
    5. `POST /verify/calls` to assert what arrived: how many times, with what
       body, answering which status.
    6. `POST /scenarios/reset` to put stateful routes back to their first step.

    ## Identifying routes without knowing ids

    Every bulk operation takes the same **selector**: `ids`, `tag`, `method`,
    `path`, `path_contains`, `response_type` or `search`. They combine with AND.
    A script that created its mocks with a tag never needs to learn the numeric
    ids the database assigned.

...

...

...

...

...

...

...

...
servers:
...
security:
  - {}
  - bearerAuth: []
...
paths:
...
  /routes:
    get:
      tags: [Routes]
      summary: List routes
      description: |
        In priority order: the first one that matches a request is the one that
        answers. Routes carrying documentation are flagged with `has_docs`;
        `include_docs` returns that documentation inline, which is the cheap way
        to find out how a set of mocks is meant to be used before touching it.
      operationId: listRoutes
      parameters:
        - name: method
          in: query
          schema: { $ref: '#/components/schemas/HttpMethod' }
        - name: response_type
          in: query
          schema: { $ref: '#/components/schemas/ResponseType' }
        - name: active
          in: query
          schema: { type: boolean }
        - name: tag
          in: query
          description: Tag name or id.
          schema: { type: string }
        - name: search
          in: query
          description: Free text over the path, the summary, the operationId and the documentation.
          schema: { type: string }
        - name: include_docs
          in: query
          description: Include each route's documentation. Off by default because it can be long.
          schema: { type: boolean }
        - name: documented
          in: query
          description: Only routes that have documentation (true) or only those that lack it (false).
          schema: { type: boolean }
      responses:
        '200':
          description: Routes in priority order
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  documented: { type: integer, description: How many of them carry documentation }
                  routes:
                    type: array
                    items: { $ref: '#/components/schemas/Route' }
    post:
      tags: [Routes]
      summary: Create a route
      description: |
        A JSON mock needs `method`, `path`, `status_code`, `response_type` and
        `response`. On a proxy route, `response` is the target URL.

        Paths under `/api` or `/mcp` are rejected: they belong to the panel and
        a mock there would never answer.
      operationId: createRoute
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RouteInput' }
            examples:
              json:
                summary: A JSON mock
                value:
                  method: get
                  path: /fake-api/orders/42
                  status_code: '200'
                  response_type: json
                  response: '{"id":42,"status":"paid"}'
                  tags: [{ name: checkout-ok }]
              proxy:
                summary: A proxy to the real backend
                value:
                  method: any
                  path: /fake-api
                  status_code: '200'
                  response_type: proxy
                  response: https://api.example.com
          multipart/form-data:
            schema: { $ref: '#/components/schemas/RouteWithFile' }
      responses:
        '201':
          description: Route created
          content:
            application/json:
              schema:
                type: object
                properties:
                  created: { type: boolean }
                  route: { $ref: '#/components/schemas/RouteDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /routes/activate:
    post:
      tags: [Routes]
      summary: Turn routes on
      description: |
        Enables everything the selector matches.

        With `exclusive: true` it also disables any other active route that
        would compete for the same request (same path, overlapping method).
        That is what makes it a scenario switch: activating the failure mock
        puts the happy-path one away instead of leaving both on and letting the
        order decide. Routes that answer something else are never touched.
      operationId: activateRoutes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/RouteSelector'
                - type: object
                  properties:
                    exclusive:
                      type: boolean
                      default: false
                      description: Disable the routes that collide with the ones being enabled.
            examples:
              scenario:
                summary: Switch the checkout mocks to the failure variant
                value: { tag: checkout-down, exclusive: true }
      responses:
        '200':
          description: State applied
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActiveResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /routes/deactivate:
    post:
      tags: [Routes]
      summary: Turn routes off
      description: |
        Disables everything the selector matches. This is how traffic falls
        through to a proxy again without deleting the mocks.
      operationId: deactivateRoutes
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RouteSelector' }
            examples:
              byTag:
                summary: Everything tagged payments
                value: { tag: payments }
              byPath:
                summary: One endpoint, whatever its id
                value: { method: get, path: /fake-api/orders }
      responses:
        '200':
          description: State applied
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActiveResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /routes/delete:
    post:
      tags: [Routes]
      summary: Delete several routes
      description: Deleting cannot be undone; prefer `/routes/deactivate` to turn things off between test cases.
      operationId: deleteRoutes
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RouteSelector' }
      responses:
        '200':
          description: Routes deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted: { type: integer }
                  routes:
                    type: array
                    items: { $ref: '#/components/schemas/RouteRef' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /routes/tags:
    post:
      tags: [Routes]
      summary: Add or remove a tag on several routes
      description: |
        `tag` is the one being applied; the selector says to which routes. To
        select by another tag use `match_tag`, so that "remove `wip` from
        everything tagged `checkout`" is expressible.

        Unknown tags are registered automatically, so they show up in the panel
        filter.
      operationId: setRoutesTags
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - type: object
                  required: [tag, action]
                  properties:
                    tag: { type: string, description: The tag being added or removed }
                    action: { type: string, enum: [add, remove] }
                    match_tag: { type: string, description: 'Select by this other tag' }
                - $ref: '#/components/schemas/RouteSelectorNoTag'
      responses:
        '200':
          description: Tag applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated: { type: integer, description: Routes that actually changed }
                  tag: { type: string }
                  action: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /routes/reorder:
    post:
      tags: [Routes]
      summary: Set the priority order
      description: |
        When several routes could answer the same request, the lowest order
        wins. Pass the ids highest priority first.

        Proxy routes are numbered in their own high range, so they stay behind
        every mock whatever position you give them.
      operationId: reorderRoutes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [order]
              properties:
                order:
                  type: array
                  items: { type: integer }
                  description: Route ids, highest priority first.
      responses:
        '200':
          description: Order applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  reordered: { type: integer }
                  order:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: integer }
                        orden: { type: integer }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /routes/try:
    post:
      tags: [Testing]
      summary: Call a route through this server
      description: |
        Sends a real request to the configured route and returns what it
        answered. It goes through the whole pipeline, so conditions, scenarios,
        templating, latency and faults apply, and the call shows up in the log
        like any other, with its trace id.

        This is the difference between "the mock is saved" and "the mock
        answers what I think it answers".
      operationId: tryRoute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path: { type: string, example: /fake-api/orders?page=2 }
                method: { type: string, default: GET }
                body: { type: string, description: 'Request body, as text' }
                headers:
                  type: array
                  items: { $ref: '#/components/schemas/HeaderRule' }
                timeout_ms: { type: integer, default: 10000 }
      responses:
        '200':
          description: What the route answered
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: integer }
                  duration_ms: { type: integer }
                  headers: { type: object, additionalProperties: true }
                  body: { description: 'Parsed when it was JSON, raw text otherwise' }
                  trace_id: { type: string, nullable: true }
        '400': { $ref: '#/components/responses/BadRequest' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /routes/usage:
    get:
      tags: [Logs]
      summary: Which routes are actually used
      description: |
        Calls, last use, errors and average duration per route, taken from the
        log. Useful to find dead mocks before cleaning up, or to confirm the
        traffic you expected actually landed. Bounded by log retention.
      operationId: routeUsage
      parameters:
        - name: since_ms
          in: query
          description: Epoch milliseconds lower bound. Without it, everything still in the log.
          schema: { type: integer, format: int64 }
        - name: include_unused
          in: query
          description: Also list routes with no calls. Default true.
          schema: { type: boolean }
      responses:
        '200':
          description: Usage per route
          content:
            application/json:
              schema:
                type: object
                properties:
                  routes:
                    type: array
                    items: { $ref: '#/components/schemas/RouteUsage' }
                  unused: { type: integer }
                  total: { type: integer }
  /routes/{id}:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    get:
      tags: [Routes]
      summary: Full detail of a route
      description: Body, headers, conditions, scenario and, for proxies, fallbacks and transforms.
      operationId: getRoute
      responses:
        '200':
          description: The route
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Routes]
      summary: Change some fields of a route
      description: |
        Only what you send is changed; everything else is kept, including the
        parts that are configured elsewhere (conditions, scenario, recording,
        latency). This is the safe way to flip one field, for example
        `{"active": false}` or `{"status_code": "500"}`.
      operationId: updateRoute
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoutePatch' }
            examples:
              disable:
                summary: Turn one route off
                value: { active: false }
              failure:
                summary: Make it answer 500
                value: { status_code: '500', response: '{"error":"boom"}' }
          multipart/form-data:
            schema: { $ref: '#/components/schemas/RouteWithFile' }
      responses:
        '200':
          description: Route updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated: { type: boolean }
                  route: { $ref: '#/components/schemas/RouteDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Routes]
      summary: Delete a route
      description: 'Deletes the route and everything attached to it: conditions, fallbacks, operations, messages.'
      operationId: deleteRoute
      responses:
        '200':
          description: Route deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted: { type: boolean }
                  id: { type: integer }
        '404': { $ref: '#/components/responses/NotFound' }
  /routes/{id}/duplicate:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    post:
      tags: [Routes]
      summary: Copy a route to another path
      description: Copies conditions, fallbacks, GraphQL operations and WebSocket messages too, which is what makes it useful for building a variant of a flow.
      operationId: duplicateRoute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_path]
              properties:
                new_path: { type: string, example: /fake-api/orders-v2 }
      responses:
        '201':
          description: Copy created
          content:
            application/json:
              schema:
                type: object
                properties:
                  created: { type: boolean }
                  route: { $ref: '#/components/schemas/RouteDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /routes/{id}/docs:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    get:
      tags: [Routes]
      summary: Read a route's documentation
      description: What the route simulates, how it is meant to be called, what to be careful with. Worth reading before changing a route somebody else configured.
      operationId: getRouteDocs
      responses:
        '200':
          description: The documentation
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  method: { type: string }
                  path: { type: string }
                  docs: { type: string, nullable: true }
                  documented: { type: boolean }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Routes]
      summary: Write a route's documentation
      description: |
        Replaces the documentation, or adds to it with `append`. Markdown is the
        convention. Only this field is touched, so nothing else about the route
        can be lost. An empty string clears it.
      operationId: setRouteDocs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [docs]
              properties:
                docs: { type: string }
                append: { type: boolean, default: false }
      responses:
        '200':
          description: Documentation saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated: { type: boolean }
                  id: { type: integer }
                  docs: { type: string, nullable: true }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ===== ROUTE BEHAVIOUR =====
...
  /routes/{id}/sequence/reset:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    post:
      tags: [Route behaviour]
      summary: Restart one route's scenario
      description: Puts the call counter back to zero, so the next call gets the first step again. Call it between test cases.
      operationId: resetRouteSequence
      responses:
        '200':
          description: Scenario reset
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScenarioReset' }
...
  /tags/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    delete:
      tags: [Tags]
      summary: Delete a tag
      description: Removes it from the registry and from every route carrying it.
      operationId: deleteTag
      responses:
        '200':
          description: Tag deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted: { type: boolean }
                  id: { type: string }

  # ===== ENVIRONMENTS =====
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Only enforced when the server runs with `MOCK_SERVER_API_AUTH=required`.
        Use `MOCK_SERVER_API_TOKEN` or a token created from the panel.
...